# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MMO Store API - Marketplace for source code and MMO resources
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with filters
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [newest, price_asc, price_desc, popular, rating]
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Products list with pagination
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductList"

    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product (admin/seller)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
        "404":
          description: Not found

    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # Orders
...
  /products/{productId}/reviews:
    get:
      operationId: listProductReviews
      tags: [reviews]
      summary: List reviews for a product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"

    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a review
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # Blog Posts
...
  /users/balance:
    post:
      operationId: addBalance
      tags: [users]
      summary: Add balance to user account (deposit)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BalanceInput"
      responses:
        "200":
          description: Balance updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /users/profile:
    patch:
      operationId: updateProfile
      tags: [users]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # Auth
...
    User:
      type: object
      required: [id, username, email, role, balance, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, admin]
        balance:
          type: number
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string

    # Categories
...
    Product:
      type: object
      required: [id, name, slug, price, categoryId, status, viewCount, soldCount, rating, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        shortDescription:
          type: ["string", "null"]
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        thumbnailUrl:
          type: ["string", "null"]
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, draft]
        featured:
          type: boolean
        viewCount:
          type: integer
        soldCount:
          type: integer
        rating:
          type: number
        reviewCount:
          type: integer
        tags:
          type: array
          items:
            type: string
        techStack:
          type: array
          items:
            type: string
        createdAt:
          type: string
    ProductDetail:
      type: object
      required: [id, name, slug, price, categoryId, status, viewCount, soldCount, rating, createdAt, reviews]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        shortDescription:
          type: ["string", "null"]
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        thumbnailUrl:
          type: ["string", "null"]
        imageUrls:
          type: array
          items:
            type: string
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        status:
          type: string
        featured:
          type: boolean
        viewCount:
          type: integer
        soldCount:
          type: integer
        rating:
          type: number
        reviewCount:
          type: integer
        tags:
          type: array
          items:
            type: string
        techStack:
          type: array
          items:
            type: string
        features:
          type: array
          items:
            type: string
        demoUrl:
          type: ["string", "null"]
        downloadUrl:
          type: ["string", "null"]
        createdAt:
          type: string
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
    ProductList:
      type: object
      required: [products, total, page, limit]
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
    ProductInput:
      type: object
      required: [name, price, categoryId]
      properties:
        name:
          type: string
        description:
          type: string
        shortDescription:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        thumbnailUrl:
          type: string
        imageUrls:
          type: array
          items:
            type: string
        categoryId:
          type: integer
        featured:
          type: boolean
        status:
          type: string
          enum: [active, inactive, draft]
        tags:
          type: array
          items:
            type: string
        techStack:
          type: array
          items:
            type: string
        features:
          type: array
          items:
            type: string
        demoUrl:
          type: string
        downloadUrl:
          type: string
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        shortDescription:
          type: string
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        thumbnailUrl:
          type: ["string", "null"]
        imageUrls:
          type: array
          items:
            type: string
        categoryId:
          type: integer
        featured:
          type: boolean
        status:
          type: string
          enum: [active, inactive, draft]
        tags:
          type: array
          items:
            type: string
        techStack:
          type: array
          items:
            type: string
        features:
          type: array
          items:
            type: string
        demoUrl:
          type: ["string", "null"]
        downloadUrl:
          type: ["string", "null"]

    # Orders
...
    Review:
      type: object
      required: [id, productId, userId, rating, createdAt]
      properties:
        id:
          type: integer
        productId:
          type: integer
        userId:
          type: integer
        username:
          type: ["string", "null"]
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: ["string", "null"]
        createdAt:
          type: string
    ReviewInput:
      type: object
      required: [rating]
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string

    # Blog Posts
...
    BalanceInput:
      type: object
      required: [amount]
      properties:
        amount:
          type: number
        method:
          type: string
          enum: [bank_transfer, momo, vnpay, zalopay]
    ProfileUpdate:
      type: object
      properties:
        username:
          type: string
        avatarUrl:
          type: ["string", "null"]